move stream capnweb helpers to example app
