+++ apps/streams-example-app/e2e/vitest/stream-capnweb.test.ts
@@ -3,8 +3,8 @@ import { describe, expect, it } from "vitest";
 import { e2eStreamPath, e2eStreamPathLabel, toStreamWebSocketUrl } from "../helpers.ts";
 import type { StreamEvent, StreamEventInput } from "~/domains/streams/engine/shared/event.ts";
 import { withStreamConnectionFromBrowser } from "../../src/lib/stream-rpc.ts";
-import { withStreamConnectionFromNode } from "~/domains/streams/engine/node/connect.ts";
-import type { WebSocketFrame } from "~/domains/streams/engine/connection.ts";
+import { withStreamConnectionFromNode } from "../../src/lib/node-stream-connection.ts";
+import type { WebSocketFrame } from "../../src/lib/stream-connection.ts";
 import { durableObjectProcessorSubscriber } from "~/domains/streams/engine/shared/callable-subscriber.ts";
 const e2eIt = process.env.STREAM_STAGING_E2E === "true" ? it : it.skip;
 const e2eItFails = process.env.STREAM_STAGING_E2E === "true" ? it.fails : it.skip;
